pyproject.toml
[tool.poetry]
name = "bento-authorization-service"
version = "0.10.3"
description = "Permissions and authorization service for the Bento platform."
authors = ["David Lougheed <david.lougheed@mail.mcgill.ca>"]
license = "LGPL-3.0-only"
readme = "README.md"
packages = [{include = "bento_authorization_service"}]

[tool.poetry.scripts]
bento_authz = "bento_authorization_service.cli:main_sync"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiodns = "^3.2.0"
aiofiles = "^24.1.0"
aiohttp = "^3.10.5"
asyncpg = "^0.30.0"
bento-lib = {extras = ["fastapi"], version = "^12.4.1"}
fastapi = {extras = ["all"], version = "^0.115.6"}
jsonschema = "^4.21.1"
pydantic = "^2.7.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic-settings = "^2.1.0"
structlog = "^25.1.0"
typing-extensions = "^4.11.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"
pytest = "^8.2.1"
flake8 = "^7.0.0"
debugpy = "^1.8.6"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
